# utils/swagger/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "ProjectVoice API",
        "description": "REST API для голосового приложения ProjectVoice с поддержкой WebRTC. API предоставляет функционал для управления пользователями, серверами, каналами, сообщениями и приглашениями.",
        "version": "1.0.0",
        "contact": {
            "name": "API Support",
            "email": "support@projectvoice.com"
        }
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "scheme": "bearer",
            "in": "header",
            "description": "JWT токен для аутентификации. Формат: Bearer {token}"
        }
    },
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/api/servers/{serverId}/roles/": {
            "get": {
                "tags": ["Roles"],
                "description": "",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": ["Roles"],
                "description": "",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "example": "any"
                                },
                                "color": {
                                    "example": "any"
                                },
                                "permissions": {
                                    "example": "any"
                                },
                                "position": {
                                    "example": "any"
                                },
                                "isHoisted": {
                                    "example": "any"
                                },
                                "isMentionable": {
                                    "example": "any"
                                },
                                "serverId": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/servers/{serverId}/roles/{roleId}": {
            "patch": {
                "tags": ["Roles"],
                "description": "",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "example": "any"
                                },
                                "color": {
                                    "example": "any"
                                },
                                "permissions": {
                                    "example": "any"
                                },
                                "position": {
                                    "example": "any"
                                },
                                "isHoisted": {
                                    "example": "any"
                                },
                                "isMentionable": {
                                    "example": "any"
                                },
                                "serverId": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": ["Roles"],
                "description": "",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "serverId": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/servers/{serverId}/roles/members/{memberId}/roles/{roleId}": {
            "post": {
                "tags": ["Roles"],
                "description": "",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "memberId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "serverId": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": ["Roles"],
                "description": "",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "memberId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "serverId": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}